# docs/api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Paadi backend — auth, pots, payments, settlements",
    "title": "Paadi API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/me/activity": {
      "get": {
        "operationId": "ActivityController_globalFeed",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "properties": {
                          "actorName": {
                            "nullable": true,
                            "type": "string"
                          },
                          "amountKobo": {
                            "nullable": true,
                            "type": "integer"
                          },
                          "headline": {
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          },
                          "meta": {
                            "additionalProperties": {},
                            "nullable": true,
                            "type": "object"
                          },
                          "occurredAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "potId": {
                            "nullable": true,
                            "type": "string"
                          },
                          "potTitle": {
                            "nullable": true,
                            "type": "string"
                          },
                          "type": {
                            "enum": [
                              "contribution_received",
                              "pot_created",
                              "pot_funded",
                              "pot_settled_bank",
                              "pot_settled_bill",
                              "pot_settled_wallet",
                              "pot_expired",
                              "reminder_sent",
                              "wallet_credit",
                              "wallet_withdrawal",
                              "wallet_settlement_in"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "type",
                          "occurredAt",
                          "headline",
                          "potId",
                          "potTitle",
                          "actorName",
                          "amountKobo",
                          "meta"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "nextCursor": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "activity"
        ]
      }
    },
...
    "/pots/{id}/activity": {
      "get": {
        "operationId": "ActivityController_potFeed",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "contributions": {
                      "properties": {
                        "collectedKobo": {
                          "type": "integer"
                        },
                        "contributorCount": {
                          "type": "integer"
                        },
                        "paidSplitCount": {
                          "type": "integer"
                        },
                        "targetKobo": {
                          "type": "integer"
                        },
                        "totalSplitCount": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "contributorCount",
                        "paidSplitCount",
                        "totalSplitCount",
                        "collectedKobo",
                        "targetKobo"
                      ],
                      "type": "object"
                    },
                    "items": {
                      "items": {
                        "properties": {
                          "actorName": {
                            "nullable": true,
                            "type": "string"
                          },
                          "amountKobo": {
                            "nullable": true,
                            "type": "integer"
                          },
                          "headline": {
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          },
                          "meta": {
                            "additionalProperties": {},
                            "nullable": true,
                            "type": "object"
                          },
                          "occurredAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "potId": {
                            "nullable": true,
                            "type": "string"
                          },
                          "potTitle": {
                            "nullable": true,
                            "type": "string"
                          },
                          "type": {
                            "enum": [
                              "contribution_received",
                              "pot_created",
                              "pot_funded",
                              "pot_settled_bank",
                              "pot_settled_bill",
                              "pot_settled_wallet",
                              "pot_expired",
                              "reminder_sent",
                              "wallet_credit",
                              "wallet_withdrawal",
                              "wallet_settlement_in"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "type",
                          "occurredAt",
                          "headline",
                          "potId",
                          "potTitle",
                          "actorName",
                          "amountKobo",
                          "meta"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "nextCursor": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor",
                    "contributions"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "activity"
        ]
      }
    },
...
  },
  "servers": [],
...
}